third_party/WebKit/LayoutTests/crypto/subtle/hmac/cloneKey-empty-expected.txt

Tests structured de-cloning of empty HMAC keys

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS key.type is "secret"
PASS typeof key.extractable is 'boolean'
PASS key.extractable is true
PASS key.algorithm.name is "HMAC"
PASS key.algorithm.length is 0
PASS key.algorithm.hash.name is "SHA-1"
PASS key.usages.toString() is "sign"
PASS successfullyParsed is true

TEST COMPLETE
